refactor: error conversion
